added some logging to shots cron

+++ shots.php
@@ -14,22 +14,37 @@ function grab_image($url,$saveto){
     curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
     curl_setopt($ch, CURLOPT_BINARYTRANSFER,1);
     $raw=curl_exec($ch);
-    $err = curl_error($ch);
-    if($err) {
+
+    $code = -1;
+    if(!($errno = curl_errno($ch))) {
+        $code = curl_getinfo($ch,CURLINFO_HTTP_CODE);
+    }
+    else {
+        echo "Failure with errno ".$errno."\n";
+        $err = curl_error($ch);
         echo $err."\n";
     }
     curl_close ($ch);
+
+    echo "Statuscode ".$code."\n";
+    if($code !== 200) {        
+        return false;
+    }
+
     if(file_exists($saveto)){
         unlink($saveto);
     }
     $fp = fopen($saveto,'x');
     fwrite($fp, $raw);
     fclose($fp);
+
+    return true;
 }
 
 $width = 280;
 $height = 480;
 $api = 'https://api.browshot.com/api/v1/simple?url=%s&instance_id=12&width='.$width.'&height='.$height.'&key=iTUywlSOoqVjEbWYLFlytNcaDJi';
 foreach($pages as $name => $url) {
+    echo "Grab ".$name." --> ".$url."\n";
     grab_image(sprintf($api,$url),"site/shots/".$name.".png");
 }
